move `splitStylesheets` into `codemods/css/split`
