+++ packages/@tailwindcss-upgrade/src/index.ts
@@ -9,10 +9,10 @@ import { formatNodes } from './codemods/css/format-nodes'
 import { linkConfigs as linkConfigsToStylesheets } from './codemods/css/link'
 import { migrate as migrateStylesheet } from './codemods/css/migrate'
 import { sortBuckets } from './codemods/css/sort-buckets'
+import { split as splitStylesheets } from './codemods/css/split'
 import { migrate as migrateTemplate } from './codemods/template/migrate'
 import { prepareConfig } from './codemods/template/prepare-config'
 import { help } from './commands/help'
-import { split as splitStylesheets } from './migrate'
 import { migrateJsConfig } from './migrate-js-config'
 import { migratePostCSSConfig } from './migrate-postcss'
 import { migratePrettierPlugin } from './migrate-prettier'
